--- a/lib/buster.js
+++ b/lib/buster.js
@@ -1,3 +1,5 @@
+require('cache-require-paths');
+
 if (typeof module === "object" && typeof require === "function") {
     var buster = module.exports = require("./buster/buster-wiring");
 }