@dominykas 2015-11-22
So far I have two problems with this

* the cache file is not cleared automagically - if I hadn't suspected this is the culprit, I'd probably wouldn't have figured it out, but after yet-another-npm-install the dep tree reshuffled and things started failing
* people who check their `node_modules` into source control might not be happy with an extra file appearing and changing (which includes hardcoded paths, which will fail on other people's machines as per above point)